refactor: ui 패키지를 view 패키지로 변경
