+++ src/main/java/view/OutputView.java
@@ -1,4 +1,4 @@
-package ui;
+package view;
 
 import java.util.List;
 import java.util.Map;
